fix(ci): address Bamboo clippy regressions
- clean up clippy findings in permission analysis code
- simplify several server and engine helper types/usages
- keep generated frontend package artifacts out of the commit

Validated with:
- cargo fmt --check
- cargo clippy --all-features -- -A deprecated -A clippy::module_inception -A clippy::incompatible_msrv -A clippy::needless_borrows_for_generic_args -A clippy::too-many-arguments -A clippy::while-let-loop -A dead-code -D warnings

+++ crates/bamboo-engine/src/runtime/runner/tool_execution/output_compressor/mod.rs
@@ -415,7 +415,7 @@ pub(super) async fn maybe_compress(
         use bamboo_compression::TokenCounter;
         let counter = bamboo_compression::TiktokenTokenCounter::default();
         let tokens = counter.count_text(&result.result);
-        if tokens > max_tool_output_tokens as u32 {
+        if tokens > max_tool_output_tokens {
             let truncated =
                 truncate_to_token_budget(&result.result, max_tool_output_tokens, &counter);
             tracing::info!(
